@@ -1,79 +1,62 @@
-# coding: utf-8
 
 from __future__ import annotations
 
 __all__: list[str] = []
 
-import os
-import sys
 import json
-import io
-import time
-import threading
+import os
 import re
-from typing import Any, Callable, Type, Protocol, Optional
-
-from typing_extensions import TypeAlias
+import sys
+import threading
+import time
+from collections.abc import Callable
+from typing import Any, Protocol, TypeAlias
 
 # package infos
 from .__meta__ import (  # noqa
-    __doc__,
     __author__,
-    __email__,
+    __contact__,
     __copyright__,
     __credits__,
-    __contact__,
+    __doc__,
+    __email__,
     __license__,
     __status__,
     __version__,
 )
 
-
-Callback: TypeAlias = Callable[[Optional[Exception], Optional[Any]], None]
+Callback: TypeAlias = Callable[[Exception | None, Any | None], None]
 
 
 class InputStream(Protocol):
-
-    def fileno(self) -> int:
-        ...
+    def fileno(self) -> int: ...
 
     @property
-    def closed(self) -> bool:
-        ...
+    def closed(self) -> bool: ...
 
-    def isatty(self) -> bool:
-        ...
+    def isatty(self) -> bool: ...
 
-    def tell(self) -> int:
-        ...
+    def tell(self) -> int: ...
 
-    def seek(self, offset: int, whence: int = os.SEEK_SET, /) -> int:
-        ...
+    def seek(self, offset: int, whence: int = os.SEEK_SET, /) -> int: ...
 
-    def readline(self) -> str:
-        ...
+    def readline(self) -> str: ...
 
-    def readlines(self) -> list[str]:
-        ...
+    def readlines(self) -> list[str]: ...
 
 
 class OutputStream(Protocol):
-
-    def fileno(self) -> int:
-        ...
+    def fileno(self) -> int: ...
 
     @property
-    def closed(self) -> bool:
-        ...
+    def closed(self) -> bool: ...
 
-    def write(self, b: str) -> int:
-        ...
+    def write(self, b: str) -> int: ...
 
-    def flush(self) -> None:
-        ...
+    def flush(self) -> None: ...
 
 
-class Spec(object):
+class Spec:
     """
     This class wraps methods that create JSON-RPC 2.0 compatible string representations of
     request, response and error objects. All methods are class members, so you might never want to
@@ -160,19 +143,19 @@ class Spec(object):
             raise RPCInvalidRequest(str(e))
 
         # start building the request string
-        req = f"{{\"jsonrpc\":\"2.0\",\"method\":\"{method}\""
+        req = f'{{"jsonrpc":"2.0","method":"{method}"'
 
         # add the id when given
         if id is not None:
             # encode string ids
             if isinstance(id, str):
                 id = json.dumps(id)
-            req += f",\"id\":{id}"
+            req += f',"id":{id}'
 
         # add parameters when given
         if params is not None:
             try:
-                req += f",\"params\":{json.dumps(params)}"
+                req += f',"params":{json.dumps(params)}'
             except Exception as e:
                 raise RPCParseError(str(e))
 
@@ -204,7 +187,7 @@ class Spec(object):
 
         # build the response string
         try:
-            res = f"{{\"jsonrpc\":\"2.0\",\"id\":{id},\"result\":{json.dumps(result)}}}"
+            res = f'{{"jsonrpc":"2.0","id":{id},"result":{json.dumps(result)}}}'
         except Exception as e:
             raise RPCParseError(str(e))
 
@@ -238,12 +221,12 @@ class Spec(object):
 
         # build the inner error data
         message = get_error(code).title  # type: ignore[union-attr]
-        err_data = f"{{\"code\":{code},\"message\":\"{message}\""
+        err_data = f'{{"code":{code},"message":"{message}"'
 
         # insert data when given
         if data is not None:
             try:
-                err_data += f",\"data\":{json.dumps(data)}}}"
+                err_data += f',"data":{json.dumps(data)}}}'
             except Exception as e:
                 raise RPCParseError(str(e))
         else:
@@ -254,12 +237,12 @@ class Spec(object):
             id = json.dumps(id)
 
         # start building the error string
-        err = f"{{\"jsonrpc\":\"2.0\",\"id\":{id},\"error\":{err_data}}}"
+        err = f'{{"jsonrpc":"2.0","id":{id},"error":{err_data}}}'
 
         return err
 
 
-class RPC(object):
+class RPC:
     """
     The main class of *jsonrpyc*. Instances of this class wrap an input stream *stdin* and an output
     stream *stdout* in order to communicate with other services. A service is not even forced to be
@@ -282,10 +265,11 @@ class RPC(object):
 
         import jsonrpyc
 
-        class MyTarget(object):
 
+        class MyTarget(object):
             def greet(self, name):
                 return f"Hi, {name}!"
+
 
         jsonrpc.RPC(MyTarget())
 
@@ -363,13 +347,13 @@ class RPC(object):
         if stdin is None:
             stdin = sys.stdin
         self.original_stdin = stdin
-        self.stdin = io.open(stdin.fileno(), "rb")
+        self.stdin = open(stdin.fileno(), "rb")
 
         # open output stream
         if stdout is None:
             stdout = sys.stdout
         self.original_stdout = stdout
-        self.stdout = io.open(stdout.fileno(), "wb")
+        self.stdout = open(stdout.fileno(), "wb")
 
         # other attributes
         self._i = -1
@@ -406,8 +390,8 @@ class RPC(object):
         *,
         callback: Callback | None = None,
         block: int = 0,
-        timeout: float | int = 0,
-        params: dict | None = None
+        timeout: float = 0,
+        params: dict | None = None,
     ) -> int:
         """
         Performs an actual remote procedure call by writing a request representation (a string) to
@@ -494,7 +478,7 @@ class RPC(object):
         # dispatch to the correct handler
         if "method" in obj:
             # request
-            #self._handle_request(obj)
+            # self._handle_request(obj)
             if self.method_handlers:
                 self._handle_method(obj)
             else:
@@ -516,10 +500,10 @@ class RPC(object):
         :return: None.
         """
         try:
-            method = req['method']
+            method = req["method"]
             if method not in self.method_handlers:
                 raise ValueError(f"No handler defined for method: {method}")
-            result = self.method_handlers[method](req['params'])
+            result = self.method_handlers[method](req["params"])
             if "id" in req:
                 res = Spec.response(req["id"], result)
                 self._write(res)
@@ -530,7 +514,6 @@ class RPC(object):
                 else:
                     err = Spec.error(req["id"], -32603, data=str(e))
                 self._write(err)
-
 
     def _handle_request(self, req: dict[str, Any]) -> None:
         """
@@ -680,7 +663,7 @@ class Watchdog(threading.Thread):
         self,
         rpc: RPC,
         name: str = "watchdog",
-        interval: float | int = 0.1,
+        interval: float = 0.1,
         daemon: bool = False,
         start: bool = True,
     ) -> None:
@@ -737,23 +720,22 @@ class Watchdog(threading.Thread):
             return
 
         while not self._stopper.is_set():
-
             # stop when stdin is closed
             if self.rpc.stdin.closed:
                 break
 
             # Keep linter happy
-            if self.rpc.original_stdin and self.rpc.original_stdin.closed:  # type: ignore[attr-defined] # noqa
+            if self.rpc.original_stdin and self.rpc.original_stdin.closed:  # type: ignore[attr-defined]
                 break
 
             try:
                 line = self.rpc.stdin.readline()
-            except IOError:
+            except OSError:
                 line = None
 
             if line:
-                decoded_line = line.decode('utf-8').strip()
-                match = re.search(r'^Content-Length:\s*([0-9]+)', decoded_line, re.IGNORECASE)
+                decoded_line = line.decode("utf-8").strip()
+                match = re.search(r"^Content-Length:\s*([0-9]+)", decoded_line, re.IGNORECASE)
                 debug = False
                 if match:
                     if debug:
@@ -762,10 +744,10 @@ class Watchdog(threading.Thread):
                     _ = self.rpc.stdin.readline()
                     if debug:
                         print(f"Grabbing {content_length} bytes from stdin")
-                    msg = b''
+                    msg = b""
                     while len(msg) != content_length:
-                        msg += self.rpc.stdin.read(content_length-len(msg))
-                    decoded_msg = msg.decode('utf-8').strip()
+                        msg += self.rpc.stdin.read(content_length - len(msg))
+                    decoded_msg = msg.decode("utf-8").strip()
                     if debug:
                         print(f"Incoming jsonrpc message: {decoded_msg}")
                     self.rpc._handle(decoded_msg)
@@ -817,10 +799,7 @@ class RPCError(Exception):
         :return: Whether *code* is a valid error code range.
         """
         return (
-            isinstance(code, tuple) and
-            len(code) == 2 and
-            all(isinstance(i, int) for i in code) and
-            code[0] <= code[1]
+            isinstance(code, tuple) and len(code) == 2 and all(isinstance(i, int) for i in code) and code[0] <= code[1]
         )
 
     def __init__(self, data: str | None = None) -> None:
@@ -838,11 +817,11 @@ class RPCError(Exception):
         return self.message
 
 
-error_map_code: dict[int, Type[RPCError]] = {}
-error_map_code_range: dict[tuple[int, int], Type[RPCError]] = {}
+error_map_code: dict[int, type[RPCError]] = {}
+error_map_code_range: dict[tuple[int, int], type[RPCError]] = {}
 
 
-def register_error(cls: Type[RPCError]) -> Type[RPCError]:
+def register_error(cls: type[RPCError]) -> type[RPCError]:
     """
     Decorator that registers a new RPC error derived from :py:class:`RPCError`. The purpose of
     error registration is to have a mapping of error codes/code ranges to error classes for faster
@@ -872,7 +851,7 @@ def register_error(cls: Type[RPCError]) -> Type[RPCError]:
     return cls
 
 
-def get_error(code: int) -> Type[RPCError]:
+def get_error(code: int) -> type[RPCError]:
     """
     Returns the RPC error class that was previously registered to *code*. A ``ValueError`` is raised
     if no error class was found for *code*.
@@ -893,7 +872,6 @@ def get_error(code: int) -> Type[RPCError]:
 
 @register_error
 class RPCParseError(RPCError):
-
     code_range = (-32700, -32700)
     code = code_range[0]
     title = "Parse error"
@@ -901,7 +879,6 @@ class RPCParseError(RPCError):
 
 @register_error
 class RPCInvalidRequest(RPCError):
-
     code_range = (-32600, -32600)
     code = code_range[0]
     title = "Invalid Request"
@@ -909,7 +886,6 @@ class RPCInvalidRequest(RPCError):
 
 @register_error
 class RPCMethodNotFound(RPCError):
-
     code_range = (-32601, -32601)
     code = code_range[0]
     title = "Method not found"
@@ -917,7 +893,6 @@ class RPCMethodNotFound(RPCError):
 
 @register_error
 class RPCInvalidParams(RPCError):
-
     code_range = (-32602, -32602)
     code = code_range[0]
     title = "Invalid params"
@@ -925,7 +900,6 @@ class RPCInvalidParams(RPCError):
 
 @register_error
 class RPCInternalError(RPCError):
-
     code_range = (-32603, -32603)
     code = code_range[0]
     title = "Internal error"
@@ -933,7 +907,6 @@ class RPCInternalError(RPCError):
 
 @register_error
 class RPCServerError(RPCError):
-
     code_range = (-32099, -32000)
     code = code_range[0]  # default code when used as is
     title = "Server error"
